feat: Add base exception classes and library config
This commit introduces the foundational components for the new common-exceptions library. It includes the tsconfig.lib.json for library setup, an enum for standardized ErrorCodes, and the base exception classes (BaseException, BaseHttpException, BaseRpcException) that will be used throughout the application for creating custom, structured errors. An empty spec file is also included as part of the initial library scaffolding.